fix: константа CITIES была объявлена в 2-х модулях: const и offers
