+++ src/components/map/map.jsx
@@ -1,9 +1,9 @@
 import React, {PureComponent} from 'react';
 import {connect} from 'react-redux';
 import leaflet from 'leaflet';
-import {MAP_TYPE} from '../../types/types';
-import {MAP_ZOOM_DEFAULT, MapType} from '../../const';
-import {getCityCenter} from '../../mocks/offers';
+import {MAP_TYPE} from '@types';
+import {MAP_ZOOM_DEFAULT, MapType} from '@const';
+import {getCityCenter} from '@utils';
 import 'leaflet/dist/leaflet.css';
 import './map.css';
