+++ src/mocks/offers.js
@@ -2,15 +2,6 @@ import {AvatarGenerator} from 'random-avatar-generator';
 import {TypesHousing} from '../const';
 import {getUUIDGenerator, uniqArray} from '../utils';
 
-export const CITIES = {
-  Paris: [48.856614, 2.3522219],
-  Cologne: [50.937531, 6.960278600000038],
-  Brussels: [50.8503396, 4.3517103],
-  Amsterdam: [52.3702157, 4.8951679],
-  Hamburg: [53.5510846, 9.99368179999999],
-  Dusseldorf: [51.2277411, 6.773455600000034],
-};
-
 const COORDINATES = {
   Paris: [
     [48.868301, 2.330252],
@@ -153,8 +144,6 @@ const getRandomOwner = () => {
     isSuper: Boolean(getRandomInteger(0, 1)),
   };
 };
-
-export const getCityCenter = (cityName) => CITIES[cityName];
 
 const getPosition = ([latitude, longitude]) => ({
   latitude,
